requirements.txt
django==4.2.11
djangorestframework==3.14.0
celery==5.3.6
drf-yasg==1.21.7
django-cors-headers==4.3.1
mysqlclient==2.2.4
django-environ==0.11.2
amqp==5.2.0
redis==5.0.3
flower==2.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
python-dotenv==1.0.1
gunicorn==21.2.0 